fix: revert changes on moving conditionals from toolkit
