+++ packages/commons/src/events/index.ts
@@ -35,5 +35,6 @@ export * from './FieldType'
 export * from './ActionType'
 export * from './FieldTypeMapping'
 export * from './Conditional'
-export * from './conditionals/conditionals'
-export * from './conditionals/validate'
+// In order to infer types, we need to export the following types along with events
+export * from '../conditionals/conditionals'
+export * from '../conditionals/validate'
